setup: add support to specify base DN
This patch add support for user to choose different base DN,
then base DN which is automatically resolved.

This change is needed, because some of the LDAP implementations,
uses more namingContexts and the first one don't have to necessary
be the one user wants to use.

This patch doesn't add support to choose this for Active Directory,
as in Active Directory resolution of base DN should work correctly.

Change-Id: Iaacd214c08b1dd75f08558ac764414d9d92dd65e
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1387254
Signed-off-by: Ondra Machacek <omachace@redhat.com>