setup: VM SSO should be enabled by default
When user is asked the question if he will be using VM SSO,
the default answer was set to 'No'. This patch changes the
default answer to 'Yes'.

The reason for this is that it's problematic to change in the
future the name of the authz provider, and the '-authz' suffix
for the authorization provider is nor really needed in many cases.

Change-Id: Ic02d36f3ee12e7e11dd7dd6f3e4b23c5ee334e60
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1462815
Signed-off-by: Ondra Machacek <omachace@redhat.com>