ldap:handle search value with asterisk
This patch handles multiple asterisks in the condition value.
Former implementation allows only single asterisk in the end of the
condition value.

A substring filter may be used to determine whether an entry contains
at least one value for a specified attribute that matches a given
substring assertion. A substring assertion is comprised of the following
components:

An optional subInitial component. If present, then this text must
appear at the beginning of a matching attribute value and must not
overlap with any text matching a subAny or subFinal component.

Zero or more subAny components. If present, then the text of each
subAny component must appear somewhere in a matching attribute value in
a manner that does not overlap with any text matching a subInitial or
subFinal component. If there are multiple subAny components, then they
must appear in a matching value in the order in which they have been
provided in the substring filter without overlapping.

An optional subFinal component. If present, then this text must
appear at the end of a matching attribute value and must not overlap
with any text matching a subInitial or subAny component.

A substring filter must have at least one subInitial, subAny, or
subFinal component.

For example, the string “(cn=John*)” represents a substring filter with
a subInitial component of “John” and no subAny or subFinal components.

The string “(cn=*John*)” represents a substring filter with a single
subAny component of “John” and no subInitial or subFinal components.

The string “(cn=*John*Doe*)” represents a substring filter with two subAny
components of “John” and “Doe” and no subInitial or subFinal components.

The string “(cn=*Doe)” represents a substring filter with a subFinal
component of “Doe” and no subInitial or subAny elements.

The string “(cn=J*o*h*n*D*o*e)” represents a substring filter with a subInitial
component of “J”, five subAny components of “o”, “h”, “n”, “D”, and “o”,
and a subFinal component of “e”.

Change-Id: Ib76665998980165d5052ede3597b771235285079
Bug-Url: https://bugzilla.redhat.com/1778471
Signed-off-by: emesika <emesika@redhat.com>
Signed-off-by: Martin Perina <mperina@redhat.com>