Fix detection of IP version available
Previously IP version has been detected by iterating all interfaces and
checking their addresses. But this could lead to the issues. For
example IPv6 address is configured on the interface, but interface is
not up, so connections using IPv6 would fail. So now we are detecting IP
version only from addresses of interfaces in up status.

Also previously we have assumed that IPv4 is always available, but that
could cause an issue when IPv4 is disabled on engine machine, but LDAP
server hostname can be resolved to both IPv4 and IPv6 addresses. So now
we are enabling IPv4 support only if IPv4 is actually configured on at
least one interfaces in status up.

Change-Id: I75cadc26eb855998e0a4dbefd7da89735f8590b2
Bug-Url: https://bugzilla.redhat.com/1896779
Bug-Url: https://bugzilla.redhat.com/1940138
Bug-Url: https://bugzilla.redhat.com/1941541
Signed-off-by: Martin Perina <mperina@redhat.com>