Don't use DNS ANY query type
When running in mixed IPv4/IPv6 environment we were fetching both A and
AAAA type records withing a single request, which was translated by JNDI
into an ANY query. DNS ANY query has been deprecated by RFC 8482 and
that's probably the reason why the number of issues around DNS
resolution of LDAP server hostnames has grown significantly inn the past
year.
This patch changes the code so we are fetching A and AAAA records in
separate queries, which shouldn't use DNS ANY query.

Change-Id: I2a6bff0500b79479705f4f3bfb8e38ec7aa0d1ba
Bug-Url: https://bugzilla.redhat.com/2010658
Signed-off-by: Martin Perina <mperina@redhat.com>